auth done

+++ src/components/todo/list.js
@@ -1,15 +1,22 @@
 import { useContext, useState, useEffect } from 'react';
 import { Toast, Badge, Pagination, Form, Row, Col } from 'react-bootstrap';
+
+
+
 import { PaginationContext } from '../../context/Pagenation';
+import { AuthContext } from '../../context/Auth';
 
 const TodoList = (props) => {
   const context = useContext(PaginationContext);
+  const authContext = useContext(AuthContext);
   const [currentPage, setCurrentPage] = useState(context.startingPage);
   const maxItems = context.itemCount;
   const [list, setList] = useState([]);
+
   useEffect(() => {
     setList(props.list);
   }, [props.list]);
+
   const numOfPages = list.length / maxItems + 1;
   const indexOfLastItem = currentPage * maxItems;
   const indexOfFirstItem = indexOfLastItem - maxItems;
@@ -56,10 +63,10 @@ const TodoList = (props) => {
               as="select"
               onChange={(e) => context.setItemCount(e.target.value)}
             >
-              <option value="2">Items per Page</option>
-              <option value="2">2</option>
-              <option value="4">4</option>
-              <option value="8">8</option>
+              <option value="3">Items per Page</option>
+              <option value="3">3</option>
+              <option value="5">5</option>
+              <option value="7">7</option>
               <option value={list.length}>All</option>
             </Form.Control>
           </Form>
@@ -71,17 +78,15 @@ const TodoList = (props) => {
               onChange={(e) => {
                 if (e.target.value === 'all') setList(props.list);
                 else {
-                  let completed = list.filter(
-                    (item) => item.complete === Boolean(e.target.value)
-                  );
+                  let completed = list.filter((item) => item.complete === Boolean(e.target.value));
                   setList(completed);
                   setCurrentPage(1);
                 }
               }}
             >
-              <option value='all'>Filter by</option>
+              <option value="all">Filter by</option>
               <option value={1}>Completed</option>
-              <option value=''>Pending</option>
+              <option value="">Pending</option>
             </Form.Control>
           </Form>
         </Col>
@@ -92,35 +97,41 @@ const TodoList = (props) => {
               onChange={(e) => {
                 context.setSortField(e.target.value);
                 if (e.target.value === 'all') setList(props.list);
-                else if(e.target.value === 'difficultyA') {
+                else if (e.target.value === 'difficultyA') {
                   setList(list.sort((a, b) => a.difficulty - b.difficulty));
                   setCurrentPage(1);
-                }
-                else if(e.target.value === 'difficultyD') {
+                } else if (e.target.value === 'difficultyD') {
                   setList(list.sort((a, b) => b.difficulty - a.difficulty));
                   setCurrentPage(1);
-                }
-                else if(e.target.value === 'complete') {
-                  setList(list.sort((a, b) => a.complete === b.complete ? 0 : a.complete ? 1 : -1));
+                } else if (e.target.value === 'complete') {
+                  setList(
+                    list.sort((a, b) =>
+                      a.complete === b.complete ? 0 : a.complete ? 1 : -1
+                    )
+                  );
                   setCurrentPage(1);
                 }
               }}
             >
-               <option value='all'>Sort by</option>
-              <option value='difficultyA'>Difficulty (Ascending)</option>
-              <option value='difficultyD'>Difficulty (Descending)</option>
-              <option value='complete'>Completion</option>
+              <option value="all">Sort by</option>
+              <option value="difficultyA">Difficulty (Ascending)</option>
+              <option value="difficultyD">Difficulty (Descending)</option>
+              <option value="complete">Completion</option>
             </Form.Control>
           </Form>
-          </Col>
+        </Col>
       </Row>
       {currentList.map((item) => (
         <Toast
           key={item._id}
           style={{ maxWidth: '100%' }}
           onClose={async () => {
-            await props.handleDelete(item);
-            await props.fetch();
+            if (authContext.user.capabilities.includes('delete')) {
+              await props.handleDelete(item);
+              await props.fetch();
+            } else {
+              alert("You don't have the permession to delete!");
+            }
           }}
         >
           <Toast.Header>
@@ -128,15 +139,19 @@ const TodoList = (props) => {
               pill
               variant={item.complete ? 'success' : 'warning'}
               onClick={async () => {
-                await props.handleComplete(item);
-                await props.fetch();
+                if (authContext.user.capabilities.includes('update')) {
+                  await props.handleComplete(item);
+                  await props.fetch();
+                } else {
+                  alert("You don't have the permession to update!");
+                }
               }}
               style={{ cursor: 'pointer' }}
             >
               {item.complete ? 'Complete' : 'Pending...'}
             </Badge>
             <strong className="mr-auto ml-4">{item.assignee}</strong>
-            </Toast.Header>
+          </Toast.Header>
           <Toast.Body>
             <h3
               className={`ml-3 ${
